Update ticket handling in VAMP communication: change ticket parameter type to uint16_t for consistency and adjust related logic in send and receive functions

+++ vamp_callbacks.h
@@ -86,7 +86,7 @@ uint8_t vamp_wsn_send(uint8_t * dst_addr, uint8_t * data, size_t len);
 /** @brief Callback para enviar TICKET
  *  @return true si se envió correctamente, false en caso contrario
  */
-bool vamp_wsn_send_ticket(uint8_t * dst_addr, uint8_t ticket);
+bool vamp_wsn_send_ticket(uint8_t * dst_addr, uint16_t ticket);
 
 /**
  * @brief Function for ASK WSN radio communication
